State machine refactoring

+++ Assets/Scripts/Enemies/States/EnemyHealState.cs
@@ -30,7 +30,7 @@ class EnemyHealState : EnemyState
         enemyHealth.Heal(healAmount);
         if (enemyHealth.HpPercentage >= neededHealthPercentage)
         {
-            context.TransitionTo((int)EnemyAI.StateOptions.MoveToPlayer);
+            context.TransitionTo(EnemyAI.StateOptions.MoveToPlayer);
         }
     }
 }
